requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
nptyping==2.5.0
numpy==1.26.3
pynrrd==1.0.0
requests==2.31.0
slicerio==0.1.8
typing_extensions==4.9.0
urllib3==2.1.0
